In the context menu restore the iTerm 1 'New' entry behaviour.
In iTerm 1,  the context menu had a single entry 'New' leading to a sub menu.
This behaved like the Profile menu in that by default it would open sessions
in a new tab,  and with the Alternate key would open sessions in a new
window.

Restore that behaviour replacing the entries 'New Window' and 'New Tab'.