Support the "Zoom" accessibility feature
Now Zoom users (System Preferences > Accessibility > Zoom) enjoy Zoom
following the iTerm2 cursor so that it always stays in the magnified
portion of the screen.

Unfortunately this does not work with ZoomText for Mac as it uses the
accessibility methods (accessibilityAttributeValue*) which seem to not
provide accurate information (more specifically,
lineBreak{Char,Index}Offsets_ are nil - when VoiceOver is running,
they are not and ZoomText for Mac works fine).