Always use the Unicode font for non-ASCII
Previously it would still use the ASCII font if the font contained the
character. But this is confusing, and seems to contradict what the option says
it will do. Furthermore, it is more useful this way, as users may prefer the
ASCII characters of one font but the Unicode characters of another font, even
if their ASCII font has those Unicode characters.

This fixes https://code.google.com/p/iterm2/issues/detail?id=1491.