Fix handling of shift+alt from Synergy
Synergy doesn't send a keyDown event with its modifierFlags
having a left or right directionality set with NSAlternateKeyMask.
Before this change, we distinguished Synergy's event by using
"modifierFlags == NSAlternateKeyMask." This logic doesn't work
when the user has both the "alt" and "shift" keys held down from
Synergy.

With this change we trigger the "left" option key actions in
PTYSession/PTYTextView whenever NSAlternateKeyMask is set and no right
directionality is set on modifierFlags (instead of testing for a left
directionality). This covers both the physical "left alt" button press
and the Synergy case. The trigger for "right" option actions are left
untouched.