Fix for setting BBEdit as default for semantic history
Without this, if a user has both BBEdit and TextMate installed and
chooses BBEdit as default for semantic history, iTerm2 would still
launch TextMate. This is because both were configured to use the
same URL scheme.
using that has the problem that when both BBEdit

(Also fix a typo in a debug message :)