Make the find bar first responder immediately upon opening
More aggressive fix for issue 3496. The original fix allowed
a possibility for enqueued events to be processed before
the text field gained focus, so that for example typing
cmd-F followed by 'a', the 'a' would go to the iTerm window.
This fix ensures that enqueued events go to the search field.