force encoding value to 32bit
NSStringEncoding is defined as NSUInteger, which is defined as
unsigned int in 32bit-env & unsigned long in 64bit-env.
However iTerm treats it as 32bit signed int, so casting 32bit signed
int to 64bit unsigned int results sign expansion.
So we need to clear high 32bit of encoding value to correctly match
encoding values with predefined constants.