Fixes issue 3773, introduced by commit 0ff995effc92fc0cfc6def5967f2825286d92d56.
In that commit, the way that an update was scheduled was to delay for MAX(0, timeout-timeSinceLastUpdate). So if an update was last scheduled .1 seconds ago and the current timeout is 0.033, then .1-.033<0 so an update is scheduled immediately. The logic behind that change was to do an update right away when going from a slow to a fast update rate. In particular, it improved fast key repeats by making the cursor move at a more regular interval (effectively forcing 1/30 sec updates).

It had an unforseen problem, though. In issue 3773, we see that the first bytes read cause this immediate update. The background thread is busy processing the next token in the queue, but before the main thread gets a chance to execute it, an update and drawRect run. This update happens a bit sooner than we'd like.

This commit increases the minimum delay to 1/60 second. That should be enough to keep the cursor's update rate reasonably consistent (between 30-60 fps) while leaving enough time to parse and execute all the tokens in the user's prompt so the spurious redraw exhibited in issue 3773 won't be noticeable.