Add ENQ Answerback String. Does not work yet.
The config option UI works in PreferencePanel.xib, in so far as the code builds and runs, and the new config option appears on the Profiles -> Terminal config screen, and the user-entered value is stored and retrieved from user preferences. But nothing is actually output in response to an ENQ.

However, in VT100Terminal.m, in "case VT100CC_ENQ:" the commented-out hard-coded value DOES work if uncommented.