Now it actually sends the string from the config file. It works good enough for some people now, but it does not yet interpret any kind of backslash or carat codes.
Some commercial terminals are no better, and most do not send any special characters in their _default_ config, and it is common to have to use a special answerback program on the host to request and read an answerback, so this is technically a useable feature already right now.

But most terminal emulators do provide a way to send escapes and linefeeds etc.