Changed all the anti-idle timer fields and references from float to double.
Tried to make doAntiIdle() consider _lastInput the way it used to, but it used to be hard coded to  "if(now>=_lastInput+60)", while I want it to use the idle period pref instead of 60. It's not working, but it's no worse than before. Builds and runs ok, just sends the anti-idle code every period regardless of keyboard activity.