Remove busted logic that caused tapped events to get remapped when the following was true: 1) Do not remap was bound to the keystroke pressed, and 2) The key panel is closed or it's open with a non-remapping action selected
The original logic was broken since 2.1.4. It was more correct. It remapped when:
1) Do not remap was bound to the keystroke pressed, and
2) They key panel is open with a non-remapping action selected

But that clause didn't make sense and was me overthinking things in 2011.

The simpler logic disables remapping only when a remapping actions is selected in an open key panel.

Furthermore, this change causes all keystrokes to be sent directly to the shortcut textfield when a remapping action is selected. Previously, that wouldn't happen if the keystroke was bound to Do Not Remap. This behavior made it difficult to change a DNR-bound key, and unnecessarily complicated this already crazy logic.

Issue 4284.