Big fat change that makes the relationships between core objects saner.
Formerly, SessionView held a strong reference to PTYSession. This was madness. Now PTYTab holds a mapping from SessionView to PTYSession (both strong references) and PTYSession strongly references SessionView. This required defining a delegate interface between PTYSession and SessionView so SessionView would not know about PTYSession's implementation details.

Formerly, autoresizing was selectively used to keep frames from the SessionView on down correct. Except when it wasn't when tmux was in use. Except when it was because you were dragging a tmux split pane. This was madness. Most of the autoresizing code in the terminal view is now explicit layout code in resizeSubviewsWithOldSize.

Formerly, PTYTextView resized its superview. This was madness. Now it asks its delegate to check if it needs to be resized from -refresh.

Formerly, PTYTextView override setFrameSize and called its superview's setFrameSize. This was madness and caused occasional crashes. It has been removed.

There was a minor cleanup of how the pasteboard is used when accessing files on it.

There were various small style improvements.

Fix a bug where a window has an imperfect frame when you drag a split pane into its own window

This is a squash of many commits. See the removeStrongReferenceFromSessionViewToSession branch for the smaller commits that went into this.