Add checkbox UI element to Profiles-Window tab
I added the checkbox with copy/paste of the previous element within the XCode UI editor and updated the properties. 

I could not find how to add the `outlet` hunk via that interface so directly added it by editing the XML. The id I used for the `outlet` node doesn't have any conflicts and is presumably safe.