Remove characters from hotkey descriptors since it was redundant.
Add setObjectsFromDictionary: to ProfileModel, iTermProfilePreferences, and iTermPreferencesBaseViewController , and use it when updating the hotkey prefs so only one notification gets posted and we don't unregister/register once per key.

Post a notification when a terminal window controller is created. This is used to calculate the siblings of a window controller before it's been saved in the iTermProfileHotKey's windowController field. Instead it gets saved in a windowControllerBeingBorn field, which is separate beacuse it's just a half-baked window controller that you can't do windowy things with yet without hitting weird problems since it e.g. has no sessions yet.

Fix bugs around auto hiding when there are multiple windows with the same hotkey.

Improve the behavior of shortcut input views. They now take precedence over window-switching keystrokes so you can bind a hotkey to cmd-opt-number.

When a hotkey is pressed the other iTermHotKey*s that have the same keypress are reported up from the carbon hotkey controller so multiple windows with the same hotkey can know about each other.

Fix a bug where the hotkey shortcut input field didn't set its stringValue when the pressed key was a hotkey elsewhere.