Add tests for -[iTermTextExtractor contentInRange:....].
Change how orphan tab fillers are found in contentInRange:. An out-of-bounds lookup in eolBlock: was removed (tabFillerOrphansOnRow:line+1, when line is the last line). This should speed up the method by not creating lots of index sets and searching the line twice (except when there are lots of tab filler orphans, which is rare).

Extend contentInRange:... to support truncating the head of its result instead of always truncating the tail when cappedAtSize>=0. This fixes a bug where prefixes for semantic history got their tails truncated when ignoring hard newlines (or if the session is over 2000 cells wide). Issue 4902.