Add "Raw coprocess command" to semantic history.
This patch adds the possibility to run a script / program
as a deferred script. This is necessary to have any interaction
back with iTerm.

One use case is for example a small script that will write text
back to iTerm to interact with the remote system while being
logged in via SSH.