Do not raise hotkey windows when switching spaces.
There's some history here.

In issue 3199, a problem was revealed where when using a pinned all-spaces open hotkey
window and you switch spaces. Another app becomes active (which the OS is
fond of doing when you switch spaces). Pressing the hotkey hides the window
instead of revealing it. I put in a fix to reactivate iTerm2 and make the
hotkey window key and ordered front on a space change.

The 3.1 hotkey window changes made this less aggressive by just doing
orderFrontRegardless on the window.

But it still doesn't make sense because pinned hotkey windows should be
allowed to order back. Why should they jump on top when you switch spaces?
It would be nice if the window was key for it to remain so but I can't
effectively detect that (the resign active notification comes after the
active space did change notification, sometimes with a significant delay,
so they can't be correlated).

This changes backs out the fixes to those bugs and lets the OS manage the
window order and key app when changing spaces.

I believe the original problem in issue 3199 has been fixed: a pinned
hotkey window that is not key does become ordered front, not hidden, when
the hotkey is pressed.