Add new applescript for hotkey windows. Commands that windows recognize now include: - is hotkey window - hotkey window profile - reveal hotkey window - hide hotkey window - toggle hotkey window
The application now understands:
- create hotkey window with profile xxx

If a hotkey window does not yet exist, you must create it. If it does exist, it
can't be created again but you can change its visibility. Recommended usage is
to iterate over the app's "windows" properties and search for the hotkey window
you care about by asking each window if it "is hotkey window" and if its
"hotkey window profile" is the one you're looking for. If none is found, use
"create hotkey window with profile xxx". Otherwise, use "reveal hotkey window".