Handle error output in coproesses.
Read from stderr in coprocesses. When it terminates, show an announcement offering to view the errors, which (if selected) are written to a tempfile and opened with your default textfile editor. You can silence errors from specific command. The run coprocess UI is extended to add a checkbox to control the sliencing for the current command.