Prepare versioning for 3.1.
Change the version number to 3.1.x. This affects only nightly builds, which will be named 3.1.x-nightly. When I build a beta release, the release_beta script will tack on .beta, so it must be run as release_beta.sh 3.1.0. That will produce 3.0.1.beta. This commit adds a Sparkle version comparator that special-cases versions ending in .beta so that a non-beta with the same first two parts will compare as newer (e.g., 3.1 is newer than 3.1.9.beta).