Detect user scroll when live scrolls occur.
A live scroll happens when you drag the scroller. It does not happen when you use the scroll wheel or keyboard shortcuts.

This fixes a bug where the GCD refresh timer fires during live scroll causing the screen to jump to the bottom. It wasn't an issue with NSTimer since the screen was never asked to redraw during live scroll because of the runloop mode.

See the thread on iterm2-discuss with the subject line: OSX 10.12.4 iTerm 3.1 beta 3 scrolling strange behavior..