Customize CPP-lru-cache for iTerm2.
Don't throw when get fails. Return nullptr instead.
Add peek() to get a key's value if it belongs to the cache without bumping it to mru.
Add erase() to remove a key.
Add exists() to check if a key exists.
Add size() to return the number of entries. Replace tabs with spaces. Remove files we don't need.