Replace the texture map with a new approach. Rather than one big atlas, keep a large number of smaller ones. Rather than accounting for glyphs one by one, expire whole atlases. This is a lot less work spent messing with reference counts because there's one reference count per "page" (née atlas) rather than per glyph. The downside is more draw calls and more PIU buffers to create. That is possibly parallelizable.
The code is a hell of a lot simpler now.

Also fixes a bug where we'd crash after closing a tab because we try to draw a 0x0 grid.

This commit is kind of half baked because there's untested code and memory leaks. This whole thing is still more researchy than I'd like.