Fix legacy scrollbar in metal with dark appearance.
There is some seriously weird stuff going on and I wouldn't be surprised if this were still broken. Overriding -scrollWheel: in PTYScrollView changes how scrollers render for some stupid reason. After flailing around for a while I found that this works but it might just be the mood my computer is in at the moment.