Perform bilinear interpolation on subpixel color models using a texture sampler.
I was inspired to change it when I noticed my code was producing the wrong results. For example, with black text on a background of color (78, 120, 133) there would be ugly brown fringes around text. I never did figure out why--looks like it was a rounding error somehwere--but this code is faster AND more correct.