Implement native notifications
The advanced setting which disables Growl is checked and if true, the
notification will be sent to NSUserNotificationCenter directly.

This controller is also set as the NSUserNotificationCenter delegate, and
implements methods to guarantee that the notification is shown and to react
accordingly if the user activates (clicks) the notification.

The delegate method growlNotificationWasClicked: is now redirected to the new
notificationWasClicked: method which does the same job of activating the
corresponding view. It is also used by the native notification center
delegate.