Cut out corners of metal view so the window will have rounded corners.
I had a hard time getting the mask right. A more theoretically correct one with
varying alpha values looks worse. I suppose that's to be expected since I'm
cheating. The content view should be drawn first and then masked by the
window's border. But the only way to do it *right* is for the window's
contentView to have a layer. Doing so breaks drawing (things randomly don't
draw, or draw incorrectly, or disappear). I think they're incompatible with
scroll views but I'm not sure. The one thing I do know is that you need to be
an Apple employee to use layers on macOS without going crazy.