Change the default top/bottom margin to 5 points.
This is going to be unpopular, and if you're reading this in anger, I'm sorry.
I think it looks good, but it messes up some people's existing setups, which is
a shame. You can change it back to 2 in advanced settings > height of top and
bottom margins in terminal panes, but doing so will disable the Metal renderer.

The reason is this:

On macOS, a Metal view can overlap its window's rounded corners. It looks
terrible, and is made especially obvious because Metal is switched on and off
all the time because MTKView is the leakiest abstraction ever.

There are three possible solutions.

1. Put the root view in a layer. This is a no-go because it almost halves the
frame rate.

2. Make the metal view's CAMetalLayer non-opaque and cut out the corners. This
is a no-go because it also almost halves the frame rate.

2. Do this.