Fix a bug where a divorced session might not be in sessionsInstance. Remove a lot of code to work around symptoms of this problem.
If you created a split of a divorced session with a keyboard shortcut and then edited the split, the new session's GUID was that of the shared profile. It was not in sessionsInstance.

This triggered some bug-workaround code in sanityCheck that added it to sessionsInstance. You could end up with many divorced sessions with the same GUID.

A very detailed analysis is in issue 6493.