Make IBOutlets weak.
They were unsafe_unretained, per tradition. But since I learned about weak
references in MRR, make all the things weak! Some day when everything is
ARCified these should become strong, which is the new model.

This fixes a crash when you edit a text field in advanced prefs and close the
window while still editing it. The trigger controller was released but then a
notification was posted to reload profiles and that referenced the trigger
controller from the advanced prefs controller.