Fix a bug where cmd-0 did not restore the original font size in tmux integration. Issue 6648
Description of the problem
--------------------------

There are three profiles:

1. The profile in the model
2. The session's originalProfile (which is a copy of the model profile from the
  last update)
3. The session's profile

When you change the font size, the model updates. That's done by
`-[TmuxController setTmuxFont:nonAsciiFont:hSpacing:vSpacing:]`. A call to
setSessionSpecificValues notices this and updates the session's originalProfile
and profile.

Pressing cmd-0 resets the font to the value in the original profile.

The profile in the model gets changed because all tmux sessions must have the
same font size. This ensures that new sessions are created with the right font
size.

Description of the solution
---------------------------
I think it was a mistake to update the model when you make a font size change.
I belive there were only a few code paths where that made any difference, and
they all had to do with restoring arrangements to create a new session. The
solution is to add a new key to the session saved arrangement that contains the
font overrides for the tmux session, which the tmux controller already had.
When setting the profile, replace the font in with the font overrides. I chose
this approach because passing the tmux controller itself into
setAddressBookEntry: was very difficult.

I tested the following cases:
- Create a new tab/window after adjusting font size
- Split a session
- Resize window

And the above when tmux is using the gateway's profile (not the tmux profile).