A most delightful hack to detect the emoji panel and adjust floating hotkey windows' levels.
Issue 6730 observes that when floating hotkey windows are at the status window
level they obscure the emoji panel.

If they are at a reasonable level then the dock obscures full-width hotkey
windows, which people love because they are human and I can't help that.

This thing detects when the emoji panel is gonna open and drops the level of
floating hotkey windows. Then it polls once a second to wait for the emoji
panel to close.

As a safeguard, it must detect it open within 1 second (configurable) before
deciding that it's actually open. If that doesn't happen then flip the level
back.

The way we test if the emoji panel is open is by looking at all the open
windows and seeing if there's a level 20 one owned by Emoji & Symbols.

I am both proud of and ashamed by this commit.