Add some polish to custom title name stuff.
Remove titleFormat from PTYSession's initializer.

Add a delegate method to provide the invocation to the session name controller.

Clean up prefs VC for session titles to udpate itself properly and show but
disable functions that aren't available.

Store custom func in user defaults as a tuple of (invocation, display name)

Make it easy to convert a tuple to/from a plist value.

Hack the API helper to convert NSNull in the root of a json dict to nil. There
can still be NSNulls that aren't at the root, but that's a less pathological
case.

Allow synchronous evaluation of an RPC to call source on each of its
dependencies.

Clean up the names of things in the session name controller.