Add iTermSwiftyString
This commit adds machinery for evaluating swifty strings (i.e., string
interpolations with a Swift-inspired syntax). The ultimate goal was to make it
easy to define a string, as for tab titles, that references function calls and
variables and updates itself when any of the variables' values changes.

This requires properly tokenizing swifty strings, evaluating their values,
calling functions synchronously (when possible, for builtins; when not, note
the function call and treat it as returning an empty string) and
asynchronously with a timeout.

Swifty strings track their dependencies: variables and not-yet-registered
functions.

The function call suggester was improved to have a swifty string suggester that
handles truncated swifty strings.