Fix a leak of VT100Tokens when they were queued. Issue 6647
The bug could be reproduced by running yes, close the session, and undo close.

This also fixes a bug where the fields of VT100Token were double freed when it got deallocated because destroyPooledObject was called after each field was released by dealloc. It was never noticed because they got nilled out and I didn't hit the queued token code path with zombies on.