Improve metal enablement behavior because of strange behavior that surfaced in a comment in issue 6902.
Do not keep trying to render a metal frame after a session has exited. This prevents infinitely retrying to draw posthumously. This was done in both renderTwoMetalFramesAndShowMetalView and drawFrameAndRemoveTemporarilyDisablementOfMetalForToken.

Remove sessionViewHideMetalViewUntilNextFrame, which was unnecessary because the caller already gets a temporary disablement token and releases it after a spin.

When drawFrameAndRemoveTemporarilyDisablementOfMetalForToken eventually finishes its async draw, change how it shows the metal view: instead of just turning on the wrapper, turning of PTYTextView's darwing, and setting the metal view's alpha to 1, also cause it to redraw the metal view immediately.