Enable fastpath underlines in the legacy renderer. Issue 7437
After years of frustration I finally figured this one out! It was trying not to set the font on the graphics context if the font hasn't changed becaues it's slow. That was wrong because the *context* changed, which is why the alpha mask wasn't getting drawn into. This absolutely sucked to debug both this time and in 2014, mostly because the CG APIs are such a bloody mess of state and implicit dependencies.