Introduce iTermLaunchExperienceController, which coordinates the various cold-launch activities to avoid spamming the user too much.
Remove check for incompatible software, which is outdated and was never very useful.

Fix a bug in the onboarding window where clicking a page indicator not adjacent to the current page would animate wrong. NSPageController lived up to my expectations, unfortunately.

Improve appVersionBeforeThisLaunch to work before initializeUserDefaults is called.

Revamp about box and link it to onboarding window.