Improve how cycles among swifty strings and variables are detected.
Construct a graph of dependencies between variables and swifty strings. Detect dependency cycles by doing a search on the graph.

Simplify attaching a swify string to a variable with its format and a variable with its evaluation. They now take care of it themselves without needing hand-holding.

Make swifty strings mutable so they can be long-lived to avoid churning references.

Add iTermVariableDesignator which serves as a comparable reference to a variable. It is used in the graph for detecting cycles since a reference to a particular variable could be made in many ways. The designator treats equivalent references as equals.

Add iTermDirectedGraph.
Add -[NSArray subarrayToIndexInclusive:].
Add -[NSString stringByAppendingVariablePathComponent:].
Make iTermVariableScope objects long-lived in PTYSession and PTYTab because they store dangling references. Short-lived scopes lose those before they can be resolved.