Many improvements to expression parsing and evaluation.
The code was originally written only with function calls in mind and then expressions and interpolated strings got added on and it turned into a bit of a mess. This commit cleanly separates parsing from evaluation. It also fixes a bunch of bugs with namepsaced functions and array literals. The suggesters work a lot better now.

- [python] add Alert class
- When parsing expressions all strings are swifty strings.
- Suggest built-in non-private functions.
- Support string literals in suggesters.
- Use iTermParsedExpression consistently instead of mixing literals and parsed expressions in the output of parsers.
- Remove the evaluateSwiftyStrings advanced pref because it only controlled whether the session's autoNameFormat was evaluated before being assigned. Double-evaluating a swifty string is madness. The advanced pref was added before session names were made into swifty strings. Happily it was off by default.
- Simplify the interface of iTermScriptFunctionCall to remove all the extraneous cruft that didn't belong.