Fix two gnarly bugs:
1. The expressions for script function calls which held the completion block
could get dealloced before getting called. This could cause them not to be run
at all. For example, a title provider that has a reference to
tab.window.currentTab.currentSession.id has enough dangling references that are
resolved late that the completion block is released before the script backing
it responds. See the mailing list thread with subject line "How to reevaluate
hook on window resize", may 10 2019 8:58 am

2. The session name controller did not rebuild refs when a title provider was
loaded after restoring a window with references to the not-yet-loaded title
provider. This caused it to be called exactly once.

This adds some extra debugging facilities. Now when you click on a variable in
the inspector it logs its outstanding references to the console.