Add support for spacing combining marks.
A spacing combining mark is a character that combines (somehow) with a preceding base character. For example, TAMIL VOWEL SIGN E. They might combine in funny ways, like to the left of the base character. This commit does more or less what Terminal does, which is to give such a character its own cell but draw it in conjunction with the preceding character, provided the predecessor is drawable non-ascii. It is guarded by the advanced pref 'Detect base unicode characters with lookup table' which is a bit of a misnomer. I am not courageous enough to turn this on for 3.3, but perhaps a later dot release can have it by default.