Fix an out-of-bounds read. I'm not quite sure how to reproduce it, but I can tell from the crash logs and disassembly that it read of the end of one of the _datas byte arrays.
The crash report with the smoking gun was 1565202147.246087595. It performed a read at offset 0xa003, which was past the end of a buffer of size 0xa000. This could happen if there was a fractional size or the canvas size doesn't match the _size. I don't know the root cause, but at least this should stem the bleeding.