Fix a bug where ascii letters that underflowed to the left got clipped. Issue 8185.
metalImagesForGlyphKey:asciiOffset:size:Scale:emoji used only the center point.
It looks like a cherrypick of fccce19 broke support for ascii underflowed added
in commit fd7d393. But it was further broken by commit 36d5661, which allows
glyphSize to be different than cellSize for Mojave. It did not account for
left-side underflow. Luckily, commit 11050db fixed a similar issue with
Iosevka, but only in the vertical dimension. Because vertical and horizontal
overflow are treated differently, this change is made to affect only the
monochrome code path. See the comment in PTYSession.m for more detail on this
aspect of the change.

This commit adapts that change to also support horizontal underflow.