Fix the frame of the hotkey window when there's another terminal window in lion fullscreen on a different desktop.
Before, we'd bump the hotkey window's frame up to the top of the screen if there was a lion fullscreen terminal on the same screen because the menu bar is normally hidden for Lion fullscreen windows. But that would only be the case if the fullscreen window in question were on the current desktop. Issue 6209, comments around Sept 2019.