Initialize _textEditingRow. Fixes a bug where editing the regex of any row but the first and then stopping editing *without making changes* caused the first row to change.
This happens because controlTextDidBeginEditing doesn't get called if the user doesn't actually make an edit but controlTextDidEndEditing is always called. To Apple's credit, the docs for controlTextDidEndEditing mention this. Plus one point for good docs, minus one million points for an API that has been broken long enough to go to war and vote.